+++ internal/protocol/errors.go
@@ -1,12 +1,14 @@
 package protocol
 
-import "errors"
+import sdkprotocol "github.com/qiuyier/Z-Courier/pkg/sdk/protocol"
 
 var (
-	ErrInvalidMagic       = errors.New("protocol: invalid magic")
-	ErrUnsupportedVersion = errors.New("protocol: unsupported version")
-	ErrBodyTooLarge       = errors.New("protocol: body too large")
-	ErrPacketTooShort     = errors.New("protocol: packet too short")
-	ErrLengthMismatch     = errors.New("protocol: packet length mismatch")
-	ErrFieldTooLarge      = errors.New("protocol: field too large")
+	ErrNilPacket          = sdkprotocol.ErrNilPacket
+	ErrInvalidMagic       = sdkprotocol.ErrInvalidMagic
+	ErrUnsupportedVersion = sdkprotocol.ErrUnsupportedVersion
+	ErrBodyTooLarge       = sdkprotocol.ErrBodyTooLarge
+	ErrPacketTooShort     = sdkprotocol.ErrPacketTooShort
+	ErrLengthMismatch     = sdkprotocol.ErrLengthMismatch
+	ErrFieldTooLarge      = sdkprotocol.ErrFieldTooLarge
+	ErrInvalidAck         = sdkprotocol.ErrInvalidAck
 )